Shift back to original name.
